Started packagin and modulisng code, Next need to finish moving functions to seperate file and calling in main
