+++ tempCharactertest_file.py
@@ -1,0 +1,7 @@
+from Characters.Character import *
+
+def main():
+    Create_Character()
+
+
+main()
